pyproject.toml
[tool.poetry]
name = "travelagenda"
version = "0.1.0"
description = ""
authors = ["Kader Miyanyedi <kadermiyanyedi@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
djangorestframework = "^3.11.0"
djangorestframework-jwt = "^1.11.0"
django-cors-headers = "^3.4.0"
djangorestframework-simplejwt = {git = "https://github.com/SimpleJWT/django-rest-framework-simplejwt"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
